[NEW] add draft spotify command
