+++ src/helpers/spotifyHelper.js
@@ -190,7 +190,7 @@ export const parseLink = async (
   return null; // if no song founded from spotify, Youtube, manual
 };
 
-export const deleteSongFromPlaylist = async (songId, client, personality) => {
+const deleteSongFromPlaylist = async (songId, client, personality) => {
   const tracks = [{ uri: songId }];
   try {
     await client.spotifyApi.removeTracksFromPlaylist(
@@ -200,5 +200,48 @@ export const deleteSongFromPlaylist = async (songId, client, personality) => {
     return personality.songSupressed;
   } catch {
     return personality.errorSupressing;
+  }
+};
+
+export const spotifyReply = async (foundLink, message, client, currentServer) => {
+  if (foundLink) {
+    const { answer, songId } = foundLink;
+    const newMessage = await message.reply(answer);
+    if (songId) await newMessage.react(currentServer.removeEmoji);
+    client.playlistCachedMessages = [
+      ...client.playlistCachedMessages,
+      { ...newMessage, songId },
+    ];
+  }
+};
+
+export const removeSpotify = async (messageReaction, PERSONALITY, client, currentServer) => {
+  const { message, emoji, users } = messageReaction;
+  const { removeEmoji } = currentServer;
+
+  const foundMessageSpotify = client.playlistCachedMessages.find(
+    // found corresponding spotify message
+    ({ id }) => id === message.id
+  );
+
+  if (
+    process.env.USE_SPOTIFY === "yes" &&
+    foundMessageSpotify &&
+    emoji.name === removeEmoji &&
+    users.cache
+      .map((user) => user.id) // if user reacting is the owner of spotify message
+      .includes(message.mentions.users.first().id)
+  ) {
+    const { songId } = foundMessageSpotify;
+
+    const result = await deleteSongFromPlaylist(
+      songId,
+      client,
+      PERSONALITY.spotify
+    );
+    client.playlistCachedMessages = client.playlistCachedMessages.filter(
+      ({ id }) => id !== message.id
+    );
+    await message.reply(result);
   }
 };
